- Refactored Car, Rental, and Factory classes
- Improved factory pattern implementation
- Refactored validation logic
- Added JavaDoc

+++ src/main/java/command/classes/RemoveCar.java
@@ -3,10 +3,20 @@ package command.classes;
 import car.CarRepository;
 import command.interfaces.Command;
 import exception.CommandException;
+import rental.RentalRepository;
 
 import java.util.Scanner;
 
+/**
+ * {@link command.interfaces.Command} to remove an existing car in the CarRepository.
+ */
 public class RemoveCar implements Command {
+    /**
+     * Executes the remove command for Car.
+     *
+     * @param args not used.
+     * @throws CommandException if user input is invalid or rental that uses the car exists.
+     */
     @Override
     public void execute(String args) throws Exception {
         Scanner scanner = new Scanner(System.in);
@@ -17,6 +27,10 @@ public class RemoveCar implements Command {
         scanner.nextLine();
         if(repository.searchById(id)==-1){
             throw new CommandException("No car found with ID: "+id);
+        }
+
+        if(RentalRepository.getInstance().searchByIdForCar(id)){
+            throw new CommandException("Please make sure to remove all rentals before removing the car!");
         }
 
         repository.removeCarById(id);
