+++ src/main/java/command/classes/RemoveCustomer.java
@@ -3,10 +3,20 @@ package command.classes;
 import command.interfaces.Command;
 import customer.CustomerRepository;
 import exception.CommandException;
+import rental.RentalRepository;
 
 import java.util.Scanner;
 
+/**
+ * {@link command.interfaces.Command} to remove an existing customer in the CustomerRepository.
+ */
 public class RemoveCustomer implements Command {
+    /**
+     * Executes the remove command for Customer.
+     *
+     * @param args not used.
+     * @throws CommandException if user input is invalid or rental that uses the customer exists.
+     */
     @Override
     public void execute(String args) throws Exception {
         Scanner scanner = new Scanner(System.in);
@@ -17,6 +27,10 @@ public class RemoveCustomer implements Command {
         scanner.nextLine();
         if(repository.searchById(id)==-1){
             throw new CommandException("No customer found with ID: "+id);
+        }
+
+        if(RentalRepository.getInstance().searchByIdForCustomer(id)){
+            throw new CommandException("Please make sure to remove all rentals before removing the customer!");
         }
 
         repository.removeCustomerById(id);
